% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfRatio1.R
\name{cdfRatio1}
\alias{cdfRatio1}
\title{Cumulative Distribution Function of the ratio of two normal variables}
\usage{
cdfRatio1(x, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{x}{a quantile value}

\item{mu1}{mean of numerator}

\item{mu2}{mean of denominator}

\item{sigma1}{standard deviation of numerator}

\item{sigma2}{standard deviation of denumerator}

\item{rho}{correlation coefficient between numerator and denominator}
}
\value{
cumulative distribution function at quantile x
}
\description{
Cumulative Distribution Function of the ratio of two normal variables
}
\examples{
cdfRatio(x = 2.3, mu1 = 5, mu2 = 2.5, sigma1 = 1, sigma2 = 0.5, rho = 0.4)
}
