% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_var.R
\name{compute_var}
\alias{compute_var}
\title{Title}
\usage{
compute_var(x, y)
}
\arguments{
\item{x}{First variable vector}

\item{y}{Second variable vector}
}
\value{
Variance-covariance matrix
}
\description{
Title
}
\examples{
x <- c(.26, 3.43, 7.38, 5.29, 3.32, 3.93, 4.97, 3.35, 4.76, 3.91)
y <- c(10.17,  6.12,  9.30, 10.79, 10.56,  8.26,  7.90,  8.93, 10.80,  9.06)
compute_var(x,y)
}
