% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpfRatio.R
\name{dpfRatio}
\alias{dpfRatio}
\title{Probability Distribution Function of the ratio of two normal variables}
\usage{
dpfRatio(q, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{q}{a quantile or vector of quantiles}

\item{mu1}{mean of numerator}

\item{mu2}{mean of denominator}

\item{sigma1}{standard deviation of numerator}

\item{sigma2}{standard deviation of denumerator}

\item{rho}{correlation coefficient between numerator and denominator}
}
\value{
distribution function at quantile q
}
\description{
Probability Distribution Function of the ratio of two normal variables
}
\examples{
dpfRatio(c(0.025,0.975) , mu1 = 5, mu2 = 2.5, sigma1 = 1, sigma2 = 0.5, rho = 0.4)
}
