% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalRatioCI.R
\name{NormalRatioCI}
\alias{NormalRatioCI}
\title{Title Computes confidence interval of the ratio of two normal variables}
\usage{
NormalRatioCI(mu1, mu2, sigma1, sigma2, rho, level = 0.95)
}
\arguments{
\item{mu1}{mean of numerator}

\item{mu2}{mean of denominator}

\item{sigma1}{standard deviation of numerator}

\item{sigma2}{standard deviation of denumerator}

\item{rho}{correlation coefficient between numerator and denominator}

\item{level}{confidence level required}
}
\value{
lower and upper limit of the ratio's confidence interval
}
\description{
Title Computes confidence interval of the ratio of two normal variables
}
\examples{
nci <- NormalRatioCI(mu1 = 5, mu2 = 2.5, sigma1 = 1, sigma2 = 0.5, rho = 0.4)

}
