% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatiosCI.R
\docType{package}
\name{RatiosCI}
\alias{RatiosCI}
\title{RatiosCI: A package to compute Confidence Interval of ratios}
\description{
This package implements methods to compute confidence intervals for ratios of two random variables
}
\section{RatiosCI functions}{

FiellerCI: computes Confidence Interval for the ratio of two random variables using Fieller's theorem
}

