% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfRatio.R
\name{cdfRatio}
\alias{cdfRatio}
\title{Cumulative Distribution Function of the ratio of two normal variables}
\usage{
cdfRatio(x, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{x}{a quantile value or a vector of quantile values}

\item{mu1}{mean of numerator}

\item{mu2}{mean of denominator}

\item{sigma1}{standard deviation of numerator}

\item{sigma2}{standard deviation of denumerator}

\item{rho}{correlation coefficient between numerator and denominator}
}
\value{
cumulative distribution function at quantile x
}
\description{
Cumulative Distribution Function of the ratio of two normal variables
}
\examples{
cdfRatio(x = 2.3, mu1 = 5, mu2 = 2.5, sigma1 = 1, sigma2 = 0.5, rho = 0.4)
curve(cdfRatio(x,mu1=1,mu2=0.5,sigma1=0.4,sigma2=0.6,rho=0.9),from=-5,to=5,ylim=c(0,1))
}
