% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratio.R
\name{compute_ratio}
\alias{compute_ratio}
\title{Title}
\usage{
compute_ratio(n, d)
}
\arguments{
\item{n}{Numerator}

\item{d}{Denominator}
}
\value{
Ratio of means
}
\description{
Title
}
\examples{
x <- rnorm(100,50,2)
y <- rnorm(60,25,3)
compute_ratio(x,y)
}
