% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiellerCI.R
\name{FiellerCI}
\alias{FiellerCI}
\title{Computes Fieller Confidence Interval}
\usage{
FiellerCI(numerator, denominator, alpha = 0.05)
}
\arguments{
\item{numerator}{Numerator}

\item{denominator}{Denominator}

\item{alpha}{Alpha risk level}
}
\value{
A vector of Lower limit of confidence interval, ratio of means, Upper limit of confidence interval
}
\description{
Computes Fieller Confidence Interval
}
\examples{
x <- c(3.26, 3.43, 7.38, 5.29, 3.32, 3.93, 4.97, 3.35, 4.76, 3.91)
y <- c(10.17,  6.12,  9.30, 10.79, 10.56,  8.26,  7.90,  8.93, 10.80,  9.06)
FiellerCI(x, y, 0.05)
}
\references{
Fieller E. C. "Some Problems in Interval Estimation." Journal of the Royal Statistical Society. Series B (Methodological) 16, no. 2 (1954): 175–85. http://www.jstor.org/stable/2984043.
}
